requirements.txt
nba_api>=1.4
pandas>=2.0
numpy>=1.24
requests>=2.31
scipy>=1.11
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
python-dotenv>=1.0
websocket-client>=1.6
aiohttp>=3.9
loguru>=0.7
